+++ Products/CMFPlone/controlpanel/tests/test_controlpanel_browser_language.py
@@ -45,7 +45,7 @@ class LanguageControlPanelFunctionalTest(unittest.TestCase):
 
     def test_language_control_panel_link(self):
         self.browser.open(
-            "%s/plone_control_panel" % self.portal_url)
+            "%s/@@overview-controlpanel" % self.portal_url)
         self.browser.getLink('Language').click()
         self.assertTrue("Language Settings" in self.browser.contents)