+++ Products/CMFPlone/controlpanel/tests/test_controlpanel_browser_markup.py
@@ -29,7 +29,7 @@ class MarkupControlPanelFunctionalTest(unittest.TestCase):
 
     def test_markup_control_panel_link(self):
         self.browser.open(
-            "%s/plone_control_panel" % self.portal_url)
+            "%s/@@overview-controlpanel" % self.portal_url)
         self.browser.getLink('Markup').click()
 
     def test_markup_control_panel_backlink(self):