+++ Products/CMFPlone/controlpanel/tests/test_controlpanel_browser_navigation.py
@@ -29,7 +29,7 @@ class NavigationControlPanelFunctionalTest(unittest.TestCase):
 
     def test_navigation_control_panel_link(self):
         self.browser.open(
-            "%s/plone_control_panel" % self.portal_url)
+            "%s/@@overview-controlpanel" % self.portal_url)
         self.browser.getLink('Navigation').click()
         self.assertTrue("Navigation Settings" in self.browser.contents)